+++ plugins/org.eclipse.n4js.utils/src/org/eclipse/n4js/utils/JsonUtils.java
@@ -15,24 +15,15 @@ import java.io.BufferedWriter;
 import java.io.FileNotFoundException;
 import java.io.IOException;
 import java.math.BigInteger;
-import java.nio.file.FileVisitOption;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.StandardOpenOption;
 import java.util.Arrays;
-import java.util.HashSet;
-import java.util.List;
-import java.util.Objects;
 import java.util.Set;
 import java.util.function.Supplier;
-import java.util.stream.Collectors;
-
-import org.eclipse.xtext.xbase.lib.Pair;
 
 import com.google.common.base.Joiner;
 import com.google.common.base.Optional;
-import com.google.common.collect.FluentIterable;
-import com.google.common.collect.Iterables;
 import com.google.common.collect.Sets;
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
@@ -214,140 +205,6 @@ public class JsonUtils {
 		try (BufferedWriter out = Files.newBufferedWriter(path, StandardOpenOption.TRUNCATE_EXISTING)) {
 			Gson gson = new GsonBuilder().setPrettyPrinting().create();
 			gson.toJson(jsonElement, out);
-		}
-	}
-
-	public static void addSourceFoldersToPackageJsonFile(Path path, String sourceContainerType, String... srcFolders)
-			throws FileNotFoundException, IOException {
-		Objects.requireNonNull(srcFolders);
-		if (srcFolders.length == 0) {
-			throw new IllegalArgumentException("no source folders given");
-		}
-		JsonElement root = loadJson(path);
-		JsonArray srcFolderArray = getOrCreateArrayDeepFailFast(root, UtilN4.PACKAGE_JSON__N4JS,
-				UtilN4.PACKAGE_JSON__SOURCES, sourceContainerType);
-		Set<String> existingSrcFolders = FluentIterable.from(srcFolderArray).filter(JsonPrimitive.class)
-				.transform(prim -> prim.getAsString()).toSet();
-		for (String srcFolder : srcFolders) {
-			if (existingSrcFolders.contains(srcFolder)) {
-				throw new IllegalStateException("package.json file already contains this source folder: " + srcFolder);
-			}
-			srcFolderArray.add(new JsonPrimitive(srcFolder));
-		}
-		saveJson(path, root);
-	}
-
-	public static void removeSourceFoldersFromPackageJsonFile(Path path, String... srcFolders)
-			throws FileNotFoundException, IOException {
-		Objects.requireNonNull(srcFolders);
-		if (srcFolders.length == 0) {
-			throw new IllegalArgumentException("no source folders given");
-		}
-		JsonElement root = loadJson(path);
-		JsonObject sourcesObject = getObjectDeep(root, UtilN4.PACKAGE_JSON__N4JS, UtilN4.PACKAGE_JSON__SOURCES);
-		if (sourcesObject == null || sourcesObject.keySet().isEmpty()) {
-			throw new IllegalStateException("package.json file does not contain any source folders");
-		}
-		Set<String> srcFoldersAsSet = Sets.newHashSet(srcFolders);
-		Set<String> pending = new HashSet<>(srcFoldersAsSet);
-		for (String sourceContainerType : sourcesObject.keySet()) {
-			JsonElement srcFoldersOfThisType = sourcesObject.get(sourceContainerType);
-			if (srcFoldersOfThisType != null && srcFoldersOfThisType.isJsonArray()) {
-				Iterables.removeIf((JsonArray) srcFoldersOfThisType, elem -> {
-					String name = elem != null && elem.isJsonPrimitive() && elem.getAsJsonPrimitive().isString()
-							? elem.getAsString()
-							: null;
-					if (name != null && srcFoldersAsSet.contains(name)) {
-						pending.remove(name);
-						return true;
-					}
-					return false;
-				});
-			}
-		}
-		if (!pending.isEmpty()) {
-			throw new IllegalStateException("package.json file does not contain the following source folders: "
-					+ Joiner.on(", ").join(pending));
-		}
-		saveJson(path, root);
-	}
-
-	public static void addDependenciesToPackageJsonFile(Path path,
-			Pair<String, String>... namesAndVersionConstraints) throws FileNotFoundException, IOException {
-		Objects.requireNonNull(namesAndVersionConstraints);
-		if (namesAndVersionConstraints.length == 0) {
-			throw new IllegalArgumentException("no dependencies given");
-		}
-		JsonElement root = loadJson(path);
-		JsonObject depsValue = getOrCreateObjectDeepFailFast(root, UtilN4.PACKAGE_JSON__DEPENDENCIES);
-		for (Pair<String, String> navc : namesAndVersionConstraints) {
-			String name = navc.getKey();
-			String versionConstraint = navc.getValue();
-			if (depsValue.keySet().contains(name)) {
-				throw new IOException("package.json file already contains a dependency to \"" + name + "\"");
-			}
-			depsValue.add(name, new JsonPrimitive(versionConstraint));
-		}
-		saveJson(path, root);
-	}
-
-	public static void removeDependenciesFromPackageJsonFile(Path path, String... namesToRemove)
-			throws FileNotFoundException, IOException {
-		Objects.requireNonNull(namesToRemove);
-		if (namesToRemove.length == 0) {
-			return;
-		}
-		JsonElement root = loadJson(path);
-		JsonObject depsValue = getObjectDeep(root, UtilN4.PACKAGE_JSON__DEPENDENCIES);
-		if (depsValue == null) {
-			throw new IOException("package.json does not contain any dependencies");
-		}
-		for (String name : namesToRemove) {
-			if (!depsValue.keySet().contains(name)) {
-				throw new IOException("package.json file does not contain a dependency to \"" + name + "\"");
-			}
-			depsValue.remove(name);
-		}
-		saveJson(path, root);
-	}
-
-	/**
-	 * Same as {@link #setVersionOfDependenciesInPackageJsonFile(Path, Set, JsonElement)}, but for all package.json
-	 * files in the entire folder tree below the given root folder.
-	 */
-	public static void setVersionOfDependenciesInAllPackageJsonFiles(Path rootFolder,
-			Set<String> namesOfDependencies, JsonElement versionToSet) throws FileNotFoundException, IOException {
-		List<Path> packageJsonFiles = Files.walk(rootFolder, FileVisitOption.FOLLOW_LINKS)
-				.filter(p -> UtilN4.PACKAGE_JSON.equals(p.getFileName().toString()))
-				.collect(Collectors.toList());
-		for (Path packageJsonFile : packageJsonFiles) {
-			setVersionOfDependenciesInPackageJsonFile(packageJsonFile, namesOfDependencies, versionToSet);
-		}
-	}
-
-	/**
-	 * Changes the version of all dependencies and devDependencies to packages with a name contained in
-	 * {@code namesOfDependencies} to the given version. Names given in the set that are not among the dependencies in
-	 * the package.json file will be ignored; dependencies in the package.json file that are not denoted by the names in
-	 * the set will remain unchanged.
-	 */
-	public static void setVersionOfDependenciesInPackageJsonFile(Path packageJsonFile,
-			Set<String> namesOfDependencies, JsonElement versionToSet) throws FileNotFoundException, IOException {
-		if (namesOfDependencies.isEmpty()) {
-			return;
-		}
-		JsonElement root = JsonUtils.loadJson(packageJsonFile);
-		boolean changed = false;
-		JsonElement deps = JsonUtils.getDeep(root, UtilN4.PACKAGE_JSON__DEPENDENCIES);
-		if (deps != null && deps.isJsonObject()) {
-			changed |= JsonUtils.changeProperties((JsonObject) deps, namesOfDependencies, versionToSet);
-		}
-		JsonElement devDeps = JsonUtils.getDeep(root, UtilN4.PACKAGE_JSON__DEV_DEPENDENCIES);
-		if (devDeps != null && devDeps.isJsonObject()) {
-			changed |= JsonUtils.changeProperties((JsonObject) devDeps, namesOfDependencies, versionToSet);
-		}
-		if (changed) {
-			JsonUtils.saveJson(packageJsonFile, root);
 		}
 	}
 }
